+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Features:
   - Test Run view: Add stability indicator -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/60)
+  - Provide public API -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/56)
 
 Bugfixes:
  - Fixed for downloading audits results from Audits Dashboard -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/51)